requirements.txt
numpy==1.17.3
opencv-python==4.1.2.30
lxml
tqdm
tensorflow==2.5.0
absl-py
easydict
matplotlib
pillow